+++ resources/views/components/application-logo.blade.php
@@ -1,11 +1,4 @@
 {{-- add setting set own application logo --}}
-@php
-    /** @var \App\Models\UploadedFile $logoFile */
-    $logoFile = null;
-    $logoId = \App\Models\Configuration::getInt(\App\Models\ConfigurationKey::APPEARANCE_APP_LOGO_ID);
-    if($logoId) {
-        $logoFile = \App\Models\UploadedFile::query()->find($logoId);
-    }
-@endphp
-<img src="{{ $logoFile?->getUrl() ?? \Illuminate\Support\Facades\Vite::asset("resources/images/logo.svg")}}"
+
+<img src="{{ \App\Models\ApplicationLogo::getUrl() }}"
      alt="Logo" {{ $attributes }}/>